chore: rewrite more component stories in storybook 5
